Cleaned up callback model of the bus and registration of callbacks - now supporting MVC async controllers.

+++ Samples/AsyncPages/Server/CommandMessageHandler.cs
@@ -16,9 +16,9 @@ namespace Server
             Thread.Sleep(TimeSpan.FromSeconds(1));
 
             if (message.Id % 2 == 0)
-                Bus.Return((int)ErrorCodes.Fail);
+                Bus.Return(ErrorCodes.Fail);
             else 
-                Bus.Return((int)ErrorCodes.None);
+                Bus.Return(ErrorCodes.None);
         }
     }
 }
